rename ubiquitous training worker vs ml worker
